refactor: クエリの生成処理を別ファイルに移動
